Add "application/vnd.api+json" support
