+++ src/Output/Copy/Serializer/ContentType/VdnApiJsonContentTypeSerializer.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace DoclerLabs\ApiClientGenerator\Output\Copy\Serializer\ContentType;
+
+class VdnApiJsonContentTypeSerializer extends AbstractJsonContentTypeSerializer
+{
+    public const MIME_TYPE = 'application/vnd.api+json';
+}
